Move react contexts under src/base
